Add support for object attributes in Iter call
This commit adds support for passing in object attributes to the Iter
callback. This PR makes a breaking change to the API since it adds
a new parameter to the callback. The option is currently supported
for GCS and Filesystem buckets only, but can be extended to other
providers on demand.

An alternative implementation would be to add a new IterWithAttrs
method and implement it only for a subset of providers. This will
avoid a breaking change and will make the support of the feature
more explicit. The downside is we need a new interface for something
that is only an option and that could be supported for all providers
in the future.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>