Add support for object upload options
I would like to be able to set the content type on uploaded objects
since some providers cannot show a preview if the content type is missing.

To facilitate this, I've extended the UploadObject method to accept a variable
list of options. The only introduction option is the content type, but it is
extensible for other options.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>